Add some extra options and start xml docs
