+++ engine/Orbit.Input/GameControllers/GameControllerConnectedEventArgs.cs
@@ -1,0 +1,11 @@
+namespace Orbit.Input;
+
+public class GameControllerConnectedEventArgs : EventArgs
+{
+    public GameControllerConnectedEventArgs(GameController gameController)
+    {
+        GameController = gameController;
+    }
+    
+    public GameController GameController { get; }
+}
