+++ engine/Orbit.Input/GameControllers/Stick.cs
@@ -1,7 +1,15 @@
 namespace Orbit.Input;
 
+/// <summary>
+/// Definition of a multi-axis component on a game pad. This could be a joystick, thumbstick or d-pad.
+/// </summary>
 public class Stick
 {
+    /// <summary>
+    /// Creates
+    /// </summary>
+    /// <param name="controller"></param>
+    /// <param name="name"></param>
     public Stick(GameController controller, string name)
     {
         XAxis = new ButtonValue<float>(controller, name, nameof(XAxis), FloatComparer.Default);
