--- engine/Orbit.Input/KeyboardOptions.cs
@@ -1,5 +1,0 @@
-namespace Orbit.Input;
-
-public partial class KeyboardOptions
-{
-}
