+++ engine/Orbit.Input/Keyboards/KeyboardOptions.cs
@@ -1,0 +1,8 @@
+namespace Orbit.Input;
+
+/// <summary>
+/// Defines the keyboard specific options used to override specific functionality.
+/// </summary>
+public partial class KeyboardOptions
+{
+}
